--- docs/chapters/jep/jep-1/pea.py
@@ -1,1 +1,0 @@
-../../../../jina/peapods/pea.py
