--- docs/chapters/jep/jep-2/pod.py
@@ -1,1 +1,0 @@
-../../../../jina/peapods/pod/__init__.py
