+++ tests/unit/peapods/test_pea.py
@@ -2,9 +2,9 @@ import pytest
 
 from jina.excepts import PeaFailToStart
 from jina.parser import set_pea_parser, set_pod_parser, set_gateway_parser
-from jina.peapods.peas.gateway_pea import GatewayPea
+from jina.peapods.peas.gateway import GatewayPea
 from jina.peapods.peas import BasePea
-from jina.peapods.pod import BasePod
+from jina.peapods.pods import BasePod
 
 
 @pytest.mark.parametrize('runtime', ['process', 'thread'])
