+++ tests/unit/peapods/test_remote.py
@@ -1,10 +1,10 @@
 import pytest
 
 from jina.parser import set_gateway_parser, set_pea_parser
-from jina.peapods.pod.gateway_pod import GatewayPod
+from jina.peapods.pods.gateway import GatewayPod
 
 if False:
-    from jina.peapods.peas.remote_pea import PeaSpawnHelper
+    from jina.peapods.peas.remote import PeaSpawnHelper
 
 
 @pytest.mark.skip
